Dev update

+++ src/Ares/Platform/WinAPI/WinMouseCodes.h
@@ -2,7 +2,7 @@
 
 namespace Ares {
 
-	enum class MouseCode : uint16_t;
+	enum class MouseCode : uint8_t;
 
 	MouseCode WinAPIMouseToMouseCode(uint32_t winApiMouse);
 	uint32_t MouseCodeToWinAPIMouse(MouseCode mouseCode);
